docs/beginner/module-06-filters-sorting/code-samples/python/requirements.txt
azure-search-documents>=11.4.0
azure-identity>=1.10.0
python-dotenv>=1.0.0
pytest>=7.0.0
requests>=2.28.0